Ah ha.. concated all URL data given to us as a zip file in to a single string.
Before if a ScrapeURL returned a zip file we broke this out in to a list and
only looked at the first element of the list. Surprise! The tvdb.xml
scraper expected to get all files in the zip file as a single document.
This is why actors and posters/fanart were not showing up in our results.