Have each 'lookup' reset the list of results to the last lookup.
Initially we had 'lookup' appending what it found to the existing list
of lookup results in the scrape_cli. This was a bit confusing and it got
really long really fast so instead we have it now reset the list of lookup
results to be just what was found in the last lookup command.