Move the methods for getting series/show data in to the series/show/ep objects
Before we had this hybrid way of invoking methods on the Scraper object
that would do the parser/xml work and then pass data in to the series/show
objects which would parse it. I decided to move all the method work
in to the series/show/episode objects so you call 'get_details()' on
an instance of one of those objects and it in turns would call the scraper
to parse some data. This feels a bit more normal to me.