Re-arrange it so that you call methods on 'Show' objects.
This is part of the 'objectify' work. We are basically moving
the methods you invoke off of the 'scraper' object and on to the
object that is retrieved by the scraper. So you call 'get_details()'
on the Show object returned by 'scraper.lookup()', etc.

Now just need to clean it up so it works as well as it used to.