Back to pre-objectification functionality.
Okay we are back to where we were for tv series before objectification.
It works as well as it did before (perhaps a little better with respect
to caching of retrieved url's) and now we call methods on objects
we get back instead of passing XML in to the parser and fiddling
the objects we get back.  Now to see if it still works for movies.