Implement function that cleans HTML tags out of matched input.
Use a subclass of the HTMLParser to strip HTML tags out of matched
strings. Before we were just passing things through but we finally
hit some test cases where we were getting HTML tags and this caused
minidom as it was finding weird tags it did not expect in the middle of
its stream.