setup.py'ify our scraper package.
We convert from a simple module to a package mainly because we want to
ship a set of known working xml scraper definitions. The plan is to be
able to specify which scraper you want by specifying the name of an
xml scraper definition that we supply.. or supply a file name of a
scraper to load.

NOTE: This changes what you import from "scrap" to "scraper".. nothing
else changes though.